# perf(front-matter): replace `replace()` with `trim()`