@timreichen 2025-02-09
superseded by https://github.com/denoland/std/pull/6393